@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.12'
+gem 'rails', '~> 4.1.1'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -9,17 +9,10 @@ gem 'rubygems-bundler'
 gem 'sqlite3'
 gem 'json'
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+gem 'sass-rails', '~> 4.0.3'
+gem 'coffee-rails'
 
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
-
-  gem 'uglifier', '>= 1.0.3'
-end
+gem 'uglifier', '>= 1.0.3'
 
 gem 'jquery-rails'
 
@@ -29,25 +22,25 @@ gem 'will_paginate', '~> 3.0.0'
 gem 'faker'
 gem 'haml'
 gem 'authlogic'
+gem 'scrypt'
 gem 'feedzirra', "~> 0.2.0.rc2"
 gem 'mysql2'
-gem 'thinking-sphinx', '~> 2.0.14'
+gem 'thinking-sphinx', '~> 3.1.1'
 gem 'nokogiri'
 gem 'rails_autolink'
 gem 'dynamic_form'
 gem 'gravatar_image_tag'
-gem "paperclip", "~> 2.7"
+gem 'paperclip', '~> 4.1.1'
 gem 'exception_notification', :require => 'exception_notifier'
 gem 'newrelic_rpm'
 
 group :test do
   gem 'rspec', '~> 2.14.0'
-  gem 'cucumber', '~> 1.2.1'
-  gem 'cucumber-rails', '~> 1.3.0', :require => false
+  gem 'cucumber', '~> 1.3.15'
+  gem 'cucumber-rails', '~> 1.4.1', :require => false
   gem 'launchy'
   gem 'rspec-rails', '~> 2.14.0'
   gem 'database_cleaner'
-  gem 'pickle'
   gem 'sham'
   gem 'machinist'
   gem 'timecop'
