@@ -1,8 +1,8 @@
 class DonationsController < ApplicationController
-  before_filter :require_admin, :except => :stats
+  before_filter :require_admin, except: :stats
 
   def index
-    @donations = Donation.all(:order => 'date DESC')
+    @donations = Donation.order('date DESC')
     @donation = Donation.new
   end
 
@@ -15,12 +15,12 @@ class DonationsController < ApplicationController
   end
 
   def create
-    @donation = Donation.new(params[:donation])
+    @donation = Donation.new(donation_params)
     if @donation.save
       flash[:notice] = "Successfully created donation."
       redirect_to donations_path
     else
-      render :action => 'new'
+      render action: 'new'
     end
   end
 
@@ -30,11 +30,11 @@ class DonationsController < ApplicationController
 
   def update
     @donation = Donation.find(params[:id])
-    if @donation.update_attributes(params[:donation])
+    if @donation.update_attributes(donation_params)
       flash[:notice] = "Successfully updated donation."
       redirect_to @donations
     else
-      render :action => 'edit'
+      render action: 'edit'
     end
   end
 
@@ -46,6 +46,12 @@ class DonationsController < ApplicationController
   end
 
   def stats
-    render :text => "#{Donation.total_this_month} #{Donation.total_this_year}"
+    render text: "#{Donation.total_this_month} #{Donation.total_this_year}"
+  end
+
+  private
+
+  def donation_params
+    params.require(:donation).permit(:amount, :email, :date)
   end
 end
