@@ -1,5 +1,5 @@
 class FeedEntriesController < ApplicationController
   def index
-    @entries = FeedEntry.paginate :page => params[:page], :order => "published_at DESC"
+    @entries = FeedEntry.paginate page: params[:page], order: "published_at DESC"
   end
 end
