@@ -2,8 +2,7 @@ require 'open-uri'
 require 'nokogiri'
 
 class Feed < ActiveRecord::Base
-  has_many :entries, :class_name => "FeedEntry"
-  attr_accessible :url
+  has_many :entries, class_name: "FeedEntry"
 
   def self.update_all()
     all.each do |feed|
@@ -22,12 +21,12 @@ class Feed < ActiveRecord::Base
         feed = Feedzirra::Feed.parse(open(feed_url).read)
         return unless feed || feed.is_a?(Fixnum)
         feed.entries.each do |entry|
-          unless FeedEntry.exists? :guid => entry.id
+          unless FeedEntry.exists? guid: entry.id
             entries << FeedEntry.create!(
               :name         => entry.title,
               :summary      => entry.summary,
               :url          => entry.url,
-              :published_at => entry.published,
+              published_at: entry.published,
               :guid         => entry.id
             )
           end
@@ -54,11 +53,11 @@ class Feed < ActiveRecord::Base
   end
 
   def last_n_entries(n)
-    entries.all(:order => "created_at DESC", :limit => n)
+    entries.order("created_at DESC").limit(n)
   end
 
   def last_entry_date
-    entries.first(:order => "created_at DESC").created_at
+    entries.order("created_at DESC").first.created_at
   end
 
   def parsed_feed
