@@ -3,28 +3,28 @@ require 'maxmind'
 class User < ActiveRecord::Base
   acts_as_authentic do |c|
     c.login_field = :email
-    c.validates_length_of_password_field_options = {:on => :update, :minimum => 4, :if => :has_no_credentials?}
-    c.validates_length_of_password_confirmation_field_options = {:on => :update, :minimum => 4, :if => :has_no_credentials?}
+    c.transition_from_crypto_providers = Authlogic::CryptoProviders::Sha512
+    c.validates_length_of_password_field_options = {on: :update, minimum: 4, if: :has_no_credentials?}
+    c.validates_length_of_password_confirmation_field_options = {on: :update, minimum: 4, if: :has_no_credentials?}
   end
 
   validates_presence_of :name
-  attr_protected :active
-  has_attached_file :photo, :processors => [:cropper], :styles => {
-    :small => "100x100#",
-    :medium => "300x300>",
-    :large => "500x500>"
+
+  has_attached_file :photo, processors: [:cropper], styles: {
+    small: "100x100#",
+    medium: "300x300>",
+    large: "500x500>"
   }
+  validates_attachment :photo,
+    content_type: { content_type: %r{\Aimage} },
+    size: { in: 0..1.megabyte }
+
   attr_accessor :crop_x, :crop_y, :crop_w, :crop_h
-  after_update :reprocess_photo, :if => :cropping?
+  after_update :reprocess_photo, if: :cropping?
 
-  has_many :events, :class_name => "EventLog"
+  has_many :events, class_name: "EventLog"
   has_and_belongs_to_many :saved_articles,
-    :join_table => "saved_articles", :class_name => "Article"
-
-  attr_accessible :name, :created_at, :email_delivery
-  attr_accessible :password, :password_confirmation, :email
-  attr_accessible :active, :location, :photo
-  attr_accessible :crop_x, :crop_y, :crop_w, :crop_h
+    join_table: "saved_articles", class_name: "Article"
 
   def cropping?
     !crop_x.blank? && !crop_y.blank? && !crop_w.blank? && !crop_h.blank?
@@ -70,9 +70,9 @@ class User < ActiveRecord::Base
   end
 
   def log_activation
-    events << EventLog.new(:email => self.email,
-                           :reason => "signup",
-                           :message => "Mode = #{self.email_delivery}, name = #{self.name}")
+    events << EventLog.new(email: self.email,
+                           reason: "signup",
+                           message: "Mode = #{self.email_delivery}, name = #{self.name}")
   end
 
   def secret_key
