@@ -23,11 +23,7 @@ Bcmets::Application.configure do
   config.action_dispatch.best_standards_support = :builtin
 
   # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
-
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
+  # config.active_record.mass_assignment_sanitizer = :strict
 
   # Do not compress assets
   config.assets.compress = false
@@ -35,9 +31,11 @@ Bcmets::Application.configure do
   # Expands the lines which load the assets
   config.assets.debug = true
 
+  config.eager_load = false
+
   $list_address = 'pete@petebevin.com'
 
-  config.action_mailer.default_url_options = { :host => "localhost:3000" }
+  config.action_mailer.default_url_options = { host: "localhost:3000" }
   config.action_mailer.delivery_method = :smtp
 
   password = nil
@@ -52,7 +50,7 @@ Bcmets::Application.configure do
     :user_name            => 'pbevin@gmail.com',
     :password             => password,
     :authentication       => 'plain',
-    :enable_starttls_auto => true
+    enable_starttls_auto: true
   }
   config.action_mailer.raise_delivery_errors = true
   config.action_mailer.perform_deliveries = true
