@@ -29,11 +29,11 @@ When /^I follow "([^\"]*)" within "([^\"]*)"$/ do |link, parent|
 end
 
 When /^I fill in "([^\"]*)" with "([^\"]*)"$/ do |field, value|
-  fill_in(field, :with => value)
+  fill_in(field, with: value)
 end
 
 When /^I fill in "([^\"]*)" for "([^\"]*)"$/ do |value, field|
-  fill_in(field, :with => value)
+  fill_in(field, with: value)
 end
 
 # Use this to fill in an entire form with data from a table. Example:
@@ -54,7 +54,7 @@ When /^I fill in the following:$/ do |fields|
 end
 
 When /^I select "([^\"]*)" from "([^\"]*)"$/ do |value, field|
-  select(value, :from => field)
+  select(value, from: field)
 end
 
 # Use this step in conjunction with Rail's datetime_select helper. For example:
@@ -73,7 +73,7 @@ end
 # When I select "November 23, 2004 11:20" as the "Preferred" date and time
 # And I select "November 25, 2004 10:30" as the "Alternative" date and time
 When /^I select "([^\"]*)" as the "([^\"]*)" date and time$/ do |datetime, datetime_label|
-  select_datetime(datetime, :from => datetime_label)
+  select_datetime(datetime, from: datetime_label)
 end
 
 # Use this step in conjunction with Rail's time_select helper. For example:
@@ -88,7 +88,7 @@ end
 # specify the name of the time on the form.  For example:
 # When I select "7:30AM" as the "Gym" time
 When /^I select "([^\"]*)" as the "([^\"]*)" time$/ do |time, time_label|
-  select_time(time, :from => time_label)
+  select_time(time, from: time_label)
 end
 
 # Use this step in conjunction with Rail's date_select helper.  For example:
@@ -101,7 +101,7 @@ end
 # you want to specify the name of the date on the form. For example:
 # When I select "April 26, 1982" as the "Date of Birth" date
 When /^I select "([^\"]*)" as the "([^\"]*)" date$/ do |date, date_label|
-  select_date(date, :from => date_label)
+  select_date(date, from: date_label)
 end
 
 When /^I check "([^\"]*)"$/ do |field|
