@@ -3,7 +3,7 @@ require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
 describe UserSessionsController do
   it "doesn't let an unconfirmed user login" do
     with_unconfirmed_user do |user|
-      post :create, :user_session => { :email => user.email, :password => "secret" }
+      post :create, user_session: { email: user.email, password: "secret" }
       controller.current_user.should be_nil
       assigns[:user_session].errors[:base].inspect.should == '["Your account is not active"]'
     end
