@@ -6,17 +6,17 @@ describe Donation do
   end
 
   it "should require all fields" do
-    Donation.make(:date => nil).should_not be_valid
-    Donation.make(:amount => nil).should_not be_valid
-    Donation.make(:email => nil).should_not be_valid
+    Donation.make(date: nil).should_not be_valid
+    Donation.make(amount: nil).should_not be_valid
+    Donation.make(email: nil).should_not be_valid
   end
 end
 
 describe Donation, "summaries" do
   before(:each) do
-    @d1 = Donation.make!(:date => DateTime.parse("December 31, 2008"), :amount => 10)
-    @d2 = Donation.make!(:date => DateTime.parse("March 15, 2009"), :amount => 20)
-    @d3 = Donation.make!(:date => DateTime.parse("April 15, 2009"), :amount => 40)
+    @d1 = Donation.make!(date: DateTime.parse("December 31, 2008"), amount: 10)
+    @d2 = Donation.make!(date: DateTime.parse("March 15, 2009"), amount: 20)
+    @d3 = Donation.make!(date: DateTime.parse("April 15, 2009"), amount: 40)
   end
 
   it "should count current month's donations" do
