@@ -5,9 +5,7 @@ describe FeedEntry do
     before(:each) do
       feed = Feed.create
       10.times do |n|
-        feed.entries.create(
-          :published_at => "2012-12-#{n}"
-        )
+        feed.entries.create published_at: Date.new(2012, 12, n+1)
       end
     end
     
