@@ -1,4 +1,3 @@
-Then /^(.+) should have an attached photo$/ do |who|
-  user = model(who)
-  user.photo.should be_present
+Then /^the user should have an attached photo$/ do
+  @user.reload.photo.should be_present
 end
